# course_work/distributed_system/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080/",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/committee/{committeeId}/distributor": {
      "post": {
        "tags": [
          "committee-controller"
        ],
        "operationId": "addDistributor",
        "parameters": [
          {
            "name": "committeeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Distributor"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/world/{worldId}": {
      "get": {
        "tags": [
          "world-controller"
        ],
        "operationId": "getWorld",
        "parameters": [
          {
            "name": "worldId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "418": {
            "description": "I'm a teapot",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/WorldResp"
                }
              }
            }
          }
        }
      }
    },
    "/world/{worldId}/nextYear": {
      "get": {
        "tags": [
          "world-controller"
        ],
        "operationId": "nextYear",
        "parameters": [
          {
            "name": "worldId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "418": {
            "description": "I'm a teapot",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/world/{worldId}/genocide": {
      "get": {
        "tags": [
          "world-controller"
        ],
        "operationId": "genocideStart",
        "parameters": [
          {
            "name": "worldId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "418": {
            "description": "I'm a teapot",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/world/{worldId}/amnesty": {
      "get": {
        "tags": [
          "world-controller"
        ],
        "operationId": "amnestyStrat",
        "parameters": [
          {
            "name": "worldId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "418": {
            "description": "I'm a teapot",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/semen": {
      "get": {
        "tags": [
          "controller-dlya-semena"
        ],
        "operationId": "getWorldId",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
    "/committee/{committeeId}": {
      "get": {
        "tags": [
          "committee-controller"
        ],
        "operationId": "getDistributionCommittee",
        "parameters": [
          {
            "name": "committeeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/DistributionCommittee"
                }
              }
            }
          }
        }
      }
    },
    "/committee/{committeeId}/{distributorID}": {
      "delete": {
        "tags": [
          "committee-controller"
        ],
        "operationId": "deleteDistributor",
        "parameters": [
          {
            "name": "distributorID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "committeeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Distributor": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "age": {
            "type": "integer",
            "format": "int32"
          },
          "mood": {
            "type": "string",
            "enum": [
              "GOOD",
              "BAD"
            ]
          },
          "distributorSkills": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DistributorSkill"
            }
          }
        }
      },
      "DistributorSkill": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "requiredScreams": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "DemonDto": {
        "type": "object",
        "properties": {
          "agesLeftInHell": {
            "type": "number"
          }
        }
      },
      "DistributionCommittee": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "estimationError": {
            "type": "number"
          },
          "distributors": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Distributor"
            }
          }
        }
      },
      "DistributionLayerDto": {
        "type": "object",
        "properties": {
          "screamsCount": {
            "type": "integer",
            "format": "int32"
          },
          "humans": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/HumanDto"
            }
          },
          "distributionCommittees": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DistributionCommittee"
            }
          }
        }
      },
      "HellDto": {
        "type": "object",
        "properties": {
          "producedScreams": {
            "type": "number"
          },
          "humans": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/HumanDto"
            }
          },
          "demons": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DemonDto"
            }
          }
        }
      },
      "HumanDto": {
        "type": "object",
        "properties": {
          "sex": {
            "type": "string",
            "enum": [
              "MALE",
              "FEMALE"
            ]
          }
        }
      },
      "RealWorldDto": {
        "type": "object",
        "properties": {
          "humans": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/HumanDto"
            }
          }
        }
      },
      "WorldResp": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "year": {
            "type": "integer",
            "format": "int32"
          },
          "realWorld": {
            "$ref": "#/components/schemas/RealWorldDto"
          },
          "distributionLayer": {
            "$ref": "#/components/schemas/DistributionLayerDto"
          },
          "hell": {
            "$ref": "#/components/schemas/HellDto"
          }
        }
      }
    }
  }
}